Move the comp to partial pages
