+++ pages/evidence/index.tsx
@@ -1,13 +1,7 @@
 import React, { ReactElement, useEffect, useState } from 'react'
 
-import { Button, Skeleton, Stack, Typography, styled } from '@mui/material'
-import { colors } from '@thebadge/ui-library'
-
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
-import { APP_URL } from '@/src/constants/common'
-import useBadgeByDisputeId from '@/src/hooks/subgraph/useBadgeByDisputeId'
-import CurationCriteriaLink from '@/src/pagePartials/badge/curate/CurationCriteriaLink'
-import BadgeEvidenceDisplay from '@/src/pagePartials/badge/curate/viewEvidence/BadgeEvidenceDisplay'
+import TBDataView from '@/src/pagePartials/evidence/TBDataView'
 import { useColorMode } from '@/src/providers/themeProvider'
 import { NextPageWithLayout } from '@/types/next'
 
@@ -27,11 +21,6 @@ type InjectedParams = {
   //the URL of JSON RPC endpoint connecting to a node on the network. It MUST be pointing to a node from the same chain identified by arbitrableChainID.
   arbitrableJsonRpcUrl: string
 }
-
-const Container = styled(Stack)(({ theme }) => ({
-  gap: theme.spacing(2),
-  padding: theme.spacing(2),
-}))
 
 /**
  * Evidence Iframe used on Kleros Court to display a link to our preview page,
@@ -87,69 +76,6 @@ const EvidenceIframe: NextPageWithLayout = () => {
         />
       </SafeSuspense>
     </>
-  )
-}
-
-function TBDataView({
-  arbitrableChainID,
-  disputeID,
-}: {
-  arbitrableChainID: number | undefined
-  disputeID: string | undefined
-}) {
-  const graphQueryResult = useBadgeByDisputeId(arbitrableChainID, disputeID)
-
-  function handleViewBadgeClick() {
-    const linkToSubmissionView = APP_URL + `/badge/preview/${graphQueryResult.data?.badge?.id}`
-    window.open(`${linkToSubmissionView}`, '_ blank')
-  }
-
-  return (
-    <Container>
-      <Container p={0}>
-        <Typography sx={{ fontSize: '14px !important' }}>
-          Submitted Badge Evidence -
-          <SafeSuspense
-            fallback={<Skeleton sx={{ margin: 'auto' }} variant={'text'} width={500} />}
-          >
-            {graphQueryResult.data?.badgeModel?.id && (
-              <CurationCriteriaLink
-                badgeModelId={graphQueryResult.data?.badgeModel?.id}
-                type="curate"
-              />
-            )}
-          </SafeSuspense>
-        </Typography>
-        <SafeSuspense>
-          {graphQueryResult.data?.badge?.id && (
-            <BadgeEvidenceDisplay badgeId={graphQueryResult.data?.badge?.id} />
-          )}
-        </SafeSuspense>
-      </Container>
-      <Container
-        sx={{
-          border: `1px solid ${colors.greyBackground}`,
-          borderRadius: 1,
-        }}
-      >
-        <Typography sx={{ fontSize: '14px !important' }}>
-          You can see more information about this dispute on TheBadge App
-        </Typography>
-        <Button
-          color="blue"
-          onClick={handleViewBadgeClick}
-          sx={{
-            borderRadius: 2,
-            textTransform: 'none',
-            width: 'fit-content',
-            fontSize: '16px !important',
-          }}
-          variant="text"
-        >
-          View details on TheBadge App
-        </Button>
-      </Container>
-    </Container>
   )
 }
